+++ packages/sdks/output/angular/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @builder.io/sdk-angular
 
+## 0.17.9
+
+### Patch Changes
+
+- 58ee59e: Fix: added lazy loading to video element
+
 ## 0.17.8
 
 ### Patch Changes